requirements.txt
openai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
moviepy
requests
schedule
python-dotenv
beautifulsoup4
feedparser
